Refactor: Modularize app and enhance budget visualization
- Organized code into separate modules: constants.py, data_loader.py, team_computation.py, team_selection.py, and visualization.py.
- Set team.py as the main entry point for the Streamlit application.

- Improved the visualizations and added player photos to the right of the field.
